Peer Modal: Refactor to use Modal
